==> 02_data_preprocessing.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

**Import necessary libraries**

import pandas as pd
import numpy as np
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import LabelEncoder, StandardScaler

df = pd.read_csv('/content/drive/MyDrive/Heart_Disease_Prediction/data/raw_data/heart_disease_uci.csv')
df.head()

# Check for Missing Values
# Identifies columns with missing values and their count
missing_values = df.isnull().sum()
print("Missing values in each column:\n", missing_values)


**Handle missing values**

#Handle missing values
# Filling numerical missing values with median
num_cols = ["trestbps", "chol", "thalch", "oldpeak"]
imputer = SimpleImputer(strategy="median")
df[num_cols] = imputer.fit_transform(df[num_cols])

# Filling categorical missing values with most frequent value
cat_cols = ["slope", "ca", "thal"]
imputer_cat = SimpleImputer(strategy="most_frequent")
df[cat_cols] = imputer_cat.fit_transform(df[cat_cols])

# Encode categorical variables
label_encoders = {}
categorical_columns = ["sex", "cp", "fbs", "restecg", "exang", "slope", "thal"]
for col in categorical_columns:
    le = LabelEncoder()
    df[col] = le.fit_transform(df[col])
    label_encoders[col] = le  # Store encoders for future use

# Feature Scaling (Standardization)
scaler = StandardScaler()
numeric_features = ["age", "trestbps", "chol", "thalch", "oldpeak"]
df[numeric_features] = scaler.fit_transform(df[numeric_features])

import os

# Define the file path
file_path = "/content/drive/MyDrive/Heart_Disease_Prediction/data/processed"

# Create directory if it doesn't exist
os.makedirs(file_path, exist_ok=True)

# Save the preprocessed dataset
df.to_csv(f"{file_path}/heart_disease_cleaned.csv", index=False)

print(f"Data preprocessing completed. Cleaned data saved successfully at {file_path}/heart_disease_cleaned.csv")
